functions_src/raw/send-notification-email-channel/requirements.txt
google-cloud-pubsub==2.8.0
python-dotenv==0.19.2
google-cloud-secret-manager==2.8.0
sendgrid==6.9.6